bitbake-metadata2doc.sh: use /bin/bash instead of /bin/sh
Things won't work when /bin/sh is a link to /bin/dash.  Dash doesn't
support passing arguments to source'd scripts, and that's what
setup-environment does:

  . ./oe-init-build-env $CWD/$1 > /dev/null

So, it's actually a bug in setup-environment, which has not been
caught so far because it is usually called from interactive shells
that are not limited as dash.

We work around that issue by explicitly interpreting
bitbake-metadata2doc.sh with bash.

Signed-off-by: Mario Domenech Goulart <mario@ossystems.com.br>